Moved methods from GameBoard to GameLogic
