+++ src/GameBoard.java
@@ -16,6 +16,7 @@ import java.util.Random;
  * Copyright: MIT
  */
 public class GameBoard extends JPanel implements IBoard {
+    GameLogic gl = new GameLogic();
     JPanel scorePanel = new JPanel();
     JPanel basePanel = new JPanel();
     JLabel scoreTxt = new JLabel("Score: ");
@@ -36,8 +37,8 @@ public class GameBoard extends JPanel implements IBoard {
     int score = 0;
 
     JLabel showScore = new JLabel("Score: " + score);
-    String bodyPart = '\u2584' + "";
-    String appleBit = '\u2058' + "";
+    String bodyPart = "\u2584";
+    String appleBit = "\u2058";
 
     javax.swing.Timer timer; // Ambition att byta till Timer
     //Timer timer;
@@ -51,18 +52,18 @@ public class GameBoard extends JPanel implements IBoard {
 
     public GameBoard(GuiHandler guiHandler) {
         this.guiHandler = guiHandler;
-        addLabels();
+        gl.addLabels(basePanel, labels, rows, columns, unitSize);
         board(guiHandler);
         snake.clear();
         createSnake();
         markStartPosition();
-        shuffleApplePosition();
+        gl.shuffleApplePosition(labels, apple, position, appleBit, rows, columns);
         setKeyBindings();
 
         ActionListener time = new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                move(direction,position);
+                gl.move(direction,position);
                 moved = true;
                 updateSnake();
             }
@@ -93,26 +94,6 @@ public class GameBoard extends JPanel implements IBoard {
         scorePanel.add(scoreTxt);
     }
 
-    public void move(char direction, Layout position) {
-        if (direction == 'U') position.row--;
-        if (direction == 'D') position.row++;
-        if (direction == 'L') position.column--;
-        if (direction == 'R') position.column++;
-    }
-
-    public void addLabels() {
-        for (int i = 0; i < rows; i++) {
-            for (int j = 0; j < columns; j++) {
-                labels[i][j] = new JLabel("", SwingConstants.CENTER);
-                labels[i][j].setBorder(new EtchedBorder()); // om detta används sätt fontsize till 20
-                labels[i][j].setMinimumSize(new Dimension(unitSize, unitSize));
-                labels[i][j].setMaximumSize(new Dimension(unitSize, unitSize));
-                labels[i][j].setPreferredSize(new Dimension(unitSize, unitSize));
-                labels[i][j].setFont(new Font("Andale Mono", Font.BOLD, 20)); // sätt till 20 om border nyttjas
-                basePanel.add(labels[i][j]);
-            }
-        }
-    }
 
     public void createSnake() {
         for (int i = 0; i < lengthOfSnake; i++) {
@@ -130,26 +111,12 @@ public class GameBoard extends JPanel implements IBoard {
         direction = 'R';
     }
 
-    public void shuffleApplePosition() {
-        labels[apple.row][apple.column].setForeground(null);
-        int row;
-        int col;
-        Random random = new Random();
-        while (true) {
-            row = random.nextInt(rows);
-            col = random.nextInt(columns);
-            if (position.row != row || position.column != col) break;
-        }
-        apple.newPos(row, col);
-        labels[apple.row][apple.column].setText(appleBit);
-        labels[apple.row][apple.column].setForeground(Color.red);
-    }
 
     public void updateSnake() {
         if (position.isEqualsTo(apple)) {
             haveEaten = true;
             addPoint();
-            shuffleApplePosition();
+            gl.shuffleApplePosition(labels, apple, position, appleBit, rows, columns);
         } else haveEaten = false;
 
         snake.add(new Layout(position));
@@ -188,7 +155,7 @@ public class GameBoard extends JPanel implements IBoard {
         snake.clear();
         createSnake();
         markStartPosition();
-        shuffleApplePosition();
+        gl.shuffleApplePosition(labels, apple, position, appleBit, rows, columns);
         timer.stop();
         score = -1;
         addPoint();
